+++ frontend/app/src/entities/branches/ui/branch-list-item/branch-default-badge.tsx
